# docs/swagger.yaml
basePath: /api
definitions:
  structs.APIResponse:
    properties:
      error:
        type: string
      message:
        type: string
      result: {}
    type: object
...
  structs.Login:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  structs.User:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      modified_at:
        type: string
      modified_by:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  structs.UserInput:
    properties:
      created_by:
        type: string
      modified_by:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
host: bookly-api-golang-production.up.railway.app
info:
  contact:
    name: Rigel Ramadhani W.
    url: https://github.com/rigelra15
  description: |-
    API untuk mengelola kategori dan buku di Bookly dengan menggunakan Golang dan PostgreSQL.
    Author: Rigel Ramadhani W. - Sanbercode Bootcamp Golang Batch 63
  title: Bookly API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Mendapatkan semua user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/structs.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.APIResponse'
      security:
      - JWT: []
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Membuat user baru
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/structs.UserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/structs.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.APIResponse'
      security:
      - JWT: []
      summary: Create new user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Menghapus user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.APIResponse'
      security:
      - JWT: []
      summary: Delete user
      tags:
      - Users
    get:
      description: Mendapatkan user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.APIResponse'
      security:
      - JWT: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Mengubah data user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/structs.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.APIResponse'
      security:
      - JWT: []
      summary: Update user
      tags:
      - Users
  /users/login:
    post:
      consumes:
      - application/json
      description: Autentikasi user untuk mendapatkan token JWT
      parameters:
      - description: User login credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/structs.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.APIResponse'
      summary: Login user
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"